bug fixes for time and dashboard queries

+++ src/main/scala/com/flight/aggregator/FlightSummary.scala
@@ -5,10 +5,10 @@ import slick.jdbc.PostgresProfile.api._
 * Flight Summary
 * Class used to store aggregated values for retrieved data for faster querying. Data is aggregated within the retrieved set of active flights
 */
-case class FlightSummary(time:Long, flight_count:Int,  average_velocity:Option[Double], average_altitude:Option[Double])
+case class FlightSummary(time:Int, flight_count:Int,  average_velocity:Option[Double], average_altitude:Option[Double])
 
 class FlightSummaryTable(tag:Tag) extends Table[FlightSummary](tag, "flight_summaries"){
-  def time = column[Long]("time")
+  def time = column[Int]("time")
   def flight_count = column[Int]("flight_count")
   def average_velocity = column[Option[Double]]("average_velocity")
   def average_altitude = column[Option[Double]]("average_altitude")
